i2c

+++ system/basic/libs/arch_bcm283x/include/arch/bcm283x/i2c.h
@@ -16,8 +16,8 @@ void     i2c_set_wait_time(uint32_t wait_time);
 void     i2c_set_stop_read(int32_t enable);
 void     i2c_putb(int32_t addr, int32_t regs, int32_t data);
 int32_t  i2c_getb(int32_t addr, int32_t regs);
-int32_t  i2c_puts(int32_t addr, int32_t regs, int32_t* pdat, int32_t size);
-int32_t  i2c_gets(int32_t addr, int32_t regs, int32_t* pdat, int32_t size);
+int32_t  i2c_puts(int32_t addr, int32_t regs, uint8_t* pdat, int32_t size);
+int32_t  i2c_gets(int32_t addr, int32_t regs, uint8_t* pdat, int32_t size);
 /*----------------------------------------------------------------------------*/
 #endif
 /*----------------------------------------------------------------------------*/
